requirements.txt
python-dotenv
python-binance
pre-commit
websockets